Fix GMT.
