+++ src/plotly/DataArrayFrom.scala
@@ -9,7 +9,7 @@ import typings.plotlyJs.mod.PlotType
 import typings.plotlyJs.plotlyJsBooleans.`false`
 import typings.plotlyJs.plotlyJsStrings.y
 
-import core.DateFormat
+import core.DateTimeGMT
 import core.metrics.*
 
 trait DataArrayFrom[A] extends (A => js.Array[Data]):
@@ -17,14 +17,14 @@ trait DataArrayFrom[A] extends (A => js.Array[Data]):
 
 object DataArrayFrom:
 
-  given history(using Performance[Interval], DateFormat[String]): DataArrayFrom[History] = m =>
+  given history(using Performance[Interval], DateTimeGMT[String]): DataArrayFrom[History] = m =>
     inline def box: Intervals => Data = v =>
       Data
         .PartialBoxPlotData()
         .setType(PlotType.box)
         .setY(v.map(_.mpb).toJSArray)
         .setHoverinfo(y)
-        .setName(s"${v.head.startTimeGMT}+08:00".ymd("fr-CA"))
+        .setName(s"${v.head.startTimeGMT}".ymd("fr-CA"))
         .setBoxpoints(`false`)
         .setLine(PartialScatterLine().setWidth(1))
